refactor: 피드백 반영 - price input 중복 유효성 검사 삭제 - 조건부 렌더링 로직 수정 - Modal 컴포넌트 분리
